requirements.txt
APScheduler==3.9.1
loguru==0.6.0
pyppeteer==1.0.2
PyYAML==6.0
requests==2.27.1
future~=0.18.2
aiohttp~=3.8.1